requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
scapy==2.6.1
apscheduler==3.10.4
aiosqlite==0.20.0
jinja2==3.1.4
itsdangerous==2.2.0
pydantic-settings==2.7.0
python-multipart==0.0.18
